Ensure Repository is saved even when no transactions took place
Previously if stores had been deleted but were pending purge at shutdown
(e.g. in the event of abnormal termination), and then the database was
shutdown without any transactions having been committed, then the
stores would be purged, but that state change would not be reflected
in the persisted Repository instance.

Also spotted an off by one error in this area.

+++ server/src/test/java/org/fuzzydb/server/BaseDatabaseTest.java
@@ -90,7 +90,7 @@ public abstract class BaseDatabaseTest {
 	public void deleteStore() {
 
 		try {
-			if (client != null |client.isConnected()) {
+			if (client != null && client.isConnected()) {
 				client.deleteStore(storeName);
 			}
 		} catch (UnknownStoreException e) {
